+++ lib/blip_ds.dart
@@ -1,15 +1,50 @@
 library blip_ds;
 
-/// Widgets - Chat
-export 'src/widgets/chat/ds_message_bubble.dart' show DSMessageBubble;
-export 'src/widgets/chat/ds_text_message_bubble.dart' show DSTextMessageBubble;
+/// Utils
+export 'src/utils/ds_utils.util.dart' show DSUtils;
+
+/// Widgets / Chat
+export 'src/widgets/chat/ds_message_bubble.widget.dart' show DSMessageBubble;
+export 'src/widgets/chat/ds_text_message_bubble.widget.dart'
+    show DSTextMessageBubble;
 export 'src/widgets/chat/audio/ds_audio_message_bubble.dart'
     show DSAudioMessageBubble;
 
-/// Enums
+/// Widgets / Texts
+export 'src/widgets/texts/ds_body_text.widget.dart' show DSBodyText;
+export 'src/widgets/texts/ds_headline_large_text.widget.dart'
+    show DSHeadlineLargeText;
+export 'src/widgets/texts/ds_headline_small_text.widget.dart'
+    show DSHeadlineSmallText;
+export 'src/widgets/texts/ds_caption_text.widget.dart' show DSCaptionText;
+export 'src/widgets/texts/ds_caption_small_text.widget.dart'
+    show DSCaptionSmallText;
+export 'src/widgets/texts/ds_button_text.widget.dart' show DSButtonText;
 
+/// Enums
 export 'src/enums/ds_border_radius.enum.dart' show DSBorderRadius;
 export 'src/enums/ds_align.enum.dart' show DSAlign;
 
-/// Themes
-export 'src/themes/system_colors.dart' show SystemColors;
+/// Themes / Colors
+export 'src/themes/colors/ds_colors.theme.dart' show DSColors;
+
+/// Themes / Texts
+export 'src/themes/texts/ds_text_theme.theme.dart' show DSTextTheme;
+
+/// Themes / Texts / Utils
+export 'src/themes/texts/utils/ds_font_families.theme.dart' show DSFontFamilies;
+export 'src/themes/texts/utils/ds_font_weights.theme.dart' show DSFontWeights;
+
+/// Themes / Texts / Styles
+export 'src/themes/texts/styles/ds_headline_large_text_style.theme.dart'
+    show DSHeadlineLargeTextStyle;
+export 'src/themes/texts/styles/ds_headline_small_text_style.theme.dart'
+    show DSHeadlineSmallTextStyle;
+export 'src/themes/texts/styles/ds_caption_small_text_style.theme.dart'
+    show DSCaptionSmallTextStyle;
+export 'src/themes/texts/styles/ds_body_text_style.theme.dart'
+    show DSBodyTextStyle;
+export 'src/themes/texts/styles/ds_button_text_style.theme.dart'
+    show DSButtonTextStyle;
+export 'src/themes/texts/styles/ds_caption_text_style.theme.dart'
+    show DSCaptionTextStyle;