+++ lib/src/utils/ds_utils.util.dart
@@ -1,0 +1,4 @@
+/// All utility constants that are used by this Design System.
+abstract class DSUtils {
+  static const packageName = 'blip_ds';
+}