+++ lib/src/widgets/chat/audio/ds_audio_message_bubble.dart
@@ -1,10 +1,10 @@
 import 'package:blip_ds/blip_ds.dart';
+import 'package:blip_ds/src/widgets/loading/ds_circle_loading.widget.dart';
+import 'package:blip_ds/src/widgets/buttons/ds_pause_button.widget.dart';
+import 'package:blip_ds/src/widgets/buttons/ds_play_button.widget.dart';
 import 'package:blip_ds/src/widgets/chat/audio/ds_audio_seek_bar.widget.dart';
 import 'package:blip_ds/src/controllers/chat/ds_audio_message_bubble.controller.dart';
 import 'package:blip_ds/src/widgets/chat/audio/ds_audio_speed_button.widget.dart';
-import 'package:blip_ds/src/widgets/ds_circle_loading.widget.dart';
-import 'package:blip_ds/src/widgets/ds_pause_button.widget.dart';
-import 'package:blip_ds/src/widgets/ds_play_button.widget.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:just_audio/just_audio.dart';
@@ -68,11 +68,11 @@ class _DSAudioMessageBubbleState extends State<DSAudioMessageBubble>
                     "x${_controller.audioSpeed.value.toString().replaceAll(RegExp(r'([.]*0)(?!.*\d)'), '')}",
                 onTap: _controller.setAudioSpeed,
                 borderColor: widget.align == DSAlign.right
-                    ? SystemColors.disabledText
-                    : SystemColors.neutralMediumSilver,
+                    ? DSColors.disabledText
+                    : DSColors.neutralMediumSilver,
                 color: widget.align == DSAlign.right
-                    ? SystemColors.neutralLightSnow
-                    : SystemColors.neutralDarkCity,
+                    ? DSColors.neutralLightSnow
+                    : DSColors.neutralDarkCity,
               ),
             ),
           ],
@@ -107,8 +107,8 @@ class _DSAudioMessageBubbleState extends State<DSAudioMessageBubble>
             top: 14.0,
             child: DSCircleLoadingWidget(
               color: widget.align == DSAlign.left
-                  ? SystemColors.neutralDarkRooftop
-                  : SystemColors.neutralLightSnow,
+                  ? DSColors.neutralDarkRooftop
+                  : DSColors.neutralLightSnow,
             ),
           );
         } else if (playing != true) {
@@ -122,8 +122,8 @@ class _DSAudioMessageBubbleState extends State<DSAudioMessageBubble>
           return DSPauseButtonWidget(
             onPressed: _controller.player.pause,
             color: widget.align == DSAlign.right
-                ? SystemColors.neutralLightSnow
-                : SystemColors.neutralDarkRooftop,
+                ? DSColors.neutralLightSnow
+                : DSColors.neutralDarkRooftop,
           );
         } else {
           return const SizedBox();