+++ lib/src/widgets/texts/ds_headline_large_text.widget.dart
@@ -1,0 +1,14 @@
+import 'package:blip_ds/blip_ds.dart';
+import 'package:blip_ds/src/widgets/texts/ds_text.widget.dart';
+
+/// A Design System's [Text] primarily used by large titles.
+///
+/// Sets [DSHeadlineLargeTextStyle] as [style] default value.
+class DSHeadlineLargeText extends DSText {
+  const DSHeadlineLargeText({
+    required super.text,
+    super.key,
+  }) : super(
+          style: const DSHeadlineLargeTextStyle(),
+        );
+}