+++ sample/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:blip_ds/blip_ds.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_lorem/flutter_lorem.dart';
-import 'package:get/get.dart';
+import 'package:sample/widgets/showcase/sample_message_bubble_showcase.dart';
+import 'package:sample/widgets/showcase/sample_text_style_showcase.dart';
 
 void main() {
   runApp(const SampleApp());
@@ -17,29 +17,15 @@ class SampleApp extends StatelessWidget {
       title: 'Blip Design System Showcase',
       theme: ThemeData(
         primarySwatch: Colors.blue,
+        textTheme: const DSTextTheme(),
       ),
-      home: HomePage(),
+      home: const HomePage(),
     );
   }
 }
 
 class HomePage extends StatelessWidget {
-  final _shorText = lorem(
-    words: 3,
-    paragraphs: 1,
-  );
-
-  final _longText = lorem(
-    words: 100,
-    paragraphs: 1,
-  );
-
-  final RxString _sampleText = RxString('');
-
-  final String _sampleAudio =
-      'https://s3.amazonaws.com/scifri-episodes/scifri20181123-episode.mp3';
-
-  HomePage({Key? key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -47,51 +33,13 @@ class HomePage extends StatelessWidget {
       appBar: AppBar(
         title: const Text('Blip Design System Showcase'),
       ),
-      bottomNavigationBar: Container(
-        color: SystemColors.neutralDarkCity,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Padding(
-              padding: const EdgeInsets.only(right: 10),
-              child: ElevatedButton(
-                onPressed: () => _sampleText.value = _shorText,
-                child: const Text('Texto Curto'),
-              ),
-            ),
-            ElevatedButton(
-              onPressed: () => _sampleText.value = _longText,
-              child: const Text('Texto Longo'),
-            ),
-          ],
-        ),
-      ),
       body: SafeArea(
-        child: Obx(
-          () => ListView(
-            reverse: true,
-            children: [
-              Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  DSTextMessageBubble(
-                    text: _sampleText.value.isNotEmpty
-                        ? _sampleText.value
-                        : _shorText,
-                    align: DSAlign.left,
-                  ),
-                  DSTextMessageBubble(
-                    text: _sampleText.value.isNotEmpty
-                        ? _sampleText.value
-                        : _shorText,
-                    align: DSAlign.right,
-                  ),
-                  DSAudioMessageBubble(align: DSAlign.left, uri: _sampleAudio),
-                  DSAudioMessageBubble(align: DSAlign.right, uri: _sampleAudio),
-                ],
-              ),
-            ],
-          ),
+        child: ListView(
+          children: [
+            SampleMessageBubbleShowcase(),
+            const Divider(color: DSColors.neutralDarkCity),
+            const SampleTextStyleShowcase(),
+          ],
         ),
       ),
     );